# worker/openapi.yaml
openapi: 3.0.3
info:
  title: E-Ink Board Worker API
  version: "1.0"
  description: |
    API for the Cloudflare Worker backing the EE02 e-ink display boards.

    Two very different kinds of clients call this API:

    - **Admin clients** (you, from a browser, curl, Postman, etc.) — everything
      under `/admin/*` plus `GET /current`. Authenticate with
      `Authorization: Bearer <api_key>`. Get a key by completing passkey
      registration (`/auth/register/*`) or by copying it from the admin UI at
      `/admin` ("Rotate API Key").
    - **Devices** (the ESP32 firmware) — `/device_config`, `/hash`,
      `/image_packed`, `/firmware_bin`. These use a different, HMAC-based
      per-request signature (`X-Device-MAC` / `X-Device-Nonce` /
      `X-Device-Signature`), not the Bearer key. Documented here for
      completeness but you won't normally call these by hand.

    Import this file into Postman/Insomnia/Swagger UI, set the `bearerAuth`
    security value to your API key, and every `/admin/*` request (including
    `GET /current`) is one click away.
servers:
  - url: https://{host}
    variables:
      host:
        default: localhost:8787
        description: Your deployed worker's hostname (or a local `wrangler dev` host:port).
security:
  - bearerAuth: []
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      description: >-
        Your account's API key, minted by /auth/register/verify,
        /auth/login/verify, or POST /admin/keys/rotate.
...
paths:
...
  /admin/me:
    get:
      summary: Get the logged-in account
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  id: { type: string }
                  display_name: { type: string, nullable: true }
                  is_superuser:
                    type: boolean
                    description: >-
                      Manually flipped in D1 by the project owner — see
                      migrations/0018_public_buckets.sql for the exact
                      command. There is no API to set this. Gates only
                      whether this account may mark a bucket it owns
                      is_public; the client uses it to decide whether to show
                      the "Make this public" checkbox at all.
                  sharing_public_key:
                    type: string
                    nullable: true
                    description: >-
                      The account's P-256 sharing public key (base64, raw
                      point), null until its first passkey login. Public by
                      design — it's the wrap target for every bucket key
                      shared with this account. Surfaced so a caller holding
                      only the API key (scripts/upload-images.mjs) can wrap a
                      new bucket's content key for this account exactly like
                      the dashboard's browser does.
    patch:
      summary: Set your display name
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                display_name: { type: string, maxLength: 40 }
              required: [display_name]
      responses:
        '200':
          description: OK
...
  /admin/me/sharing-key:
    patch:
      summary: Backfill this credential's PRF-wrapped sharing key
      description: >-
        See /auth/login/verify's description for why this is a separate,
        normally-authenticated endpoint rather than a field on that ceremony.
        A no-op (still 200) if this credential already has a wrap on file —
        guarded server-side, not by the client skipping the call.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                credential_id: { type: string, description: The WebAuthn credential this PRF output came from. }
                sharing_public_key: { type: string }
                wrapped_sharing_key: { type: string }
                wrap_nonce: { type: string }
              required: [credential_id, sharing_public_key, wrapped_sharing_key, wrap_nonce]
      responses:
        '200': { description: OK (written, or already had a wrap on file) }
        '400': { description: Missing a required field }
        '404': { description: Unknown credential, or not this account's }
...